Add libdxwifi python bindings and default params
The default parameters and some of the enum definitions are not
currently required and can be considered for removal in the future if
not necessary. Keeping them for now; the work was done in service of a
different approach from the one we ultimately used, and they may be
useful in the near future.

We definitely want to drop the precompiled .so file ASAP.