Remove direct usage of oresat-configs
NodeId in oresat-configs is deprecated, replaced by Card and cards.csv.
It turns out though that this project doesn't use NodeId anymore. The
import was a relic and further the import was the only usage of
oresat-configs. That means any direct reference to it can be dropped
both in pyproject.toml/requirements.txt and in the README install
instructions. It's now brought in only indirectly via oresat-olaf.